requirements.txt
alembic==1.3.0
appdirs==1.4.3
attrs==19.3.0
Babel==2.7.0
black==19.10b0
blinker==1.4
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dominate==2.4.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
pathspec==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
regex==2019.11.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.10
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.7
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
